backend/requirements.txt
baidu_aip==4.16.13
dashscope==1.18.0
Flask==3.0.3
Flask_Cors==4.0.0
Flask_RESTful==0.3.10
flask_sqlalchemy==3.1.1
numpy
onnxruntime
opencv_python
Pillow
ppocr_onnx==0.0.3.9
Requests
psycopg2-binary
beautifulsoup4
apscheduler
flask_jwt_extended
passlib
openai
chardet
onnx
soundfile
scipy